added DateRange class
