+++ spec/spec_helper.rb
@@ -16,3 +16,4 @@ require_relative "../lib/reservation"
 require_relative "../lib/reservation_manager"
 require_relative "../lib/room"
 require_relative "../lib/block"
+require_relative "../lib/date_range"
